rework from_pandoc architecture
